build: Allow building for various board configs
The board is now fully defined from its config header. make should be
provided with the board name using BOARD=myboard. It then get the MCU
type from it, and automatically add it as -include to the CFLAGS.

The board can also provide a some custom init code, which is used on
the Arduino Micro to workaround a bootload bug.